Perform statistical test to check for DE between final stages of every lineage.

Perform statistical test to check for DE between final stages of every lineage.

diffEndTest(models, ...)

# S4 method for SingleCellExperiment
diffEndTest(models, global = TRUE,
  pairwise = FALSE)

# S4 method for list
diffEndTest(models, global = TRUE, pairwise = FALSE)

Arguments

models
Typically the output from fitGAM, either a list of fitted GAM models, or an object of SingleCellExperiment class.

...
parameters including:

global
If TRUE, test for all pairwise comparisons simultaneously.

pairwise
If TRUE, test for all pairwise comparisons independently.

Value

A matrix with the wald statistic, the number of df and the p-value associated with each gene for all the tests performed. If the testing procedure was unsuccessful, the procedure will return NA test statistics and p-values.

Examples

data(gamList, package = "tradeSeq")
diffEndTest(gamList, global = TRUE, pairwise = TRUE)

#> Only two lineages; skipping pairwise comparison.

#>                   waldStat df    pvalue
#> 1700012B09Rik 5.952271e-17  1 1.0000000
#> 1700017B05Rik 1.195657e+00  1 0.2741914
#> 1700020L24Rik 5.952271e-17  1 1.0000000
#> 1810009A15Rik 7.361088e-02  1 0.7861501
#> 1810011H11Rik 5.952271e-17  1 1.0000000
#> 1810027O10Rik 6.182099e-01  1 0.4317133
#> 1810031K17Rik 8.649102e-01  1 0.3523682
#> 1810033B17Rik 5.952271e-17  1 1.0000000
#> 1810037I17Rik 1.311039e+00  1 0.2522066
#> 2010002N04Rik 5.952271e-17  1 1.0000000
#> 2010011I20Rik 2.664639e-02  1 0.8703317