Evaluate the optimal number of knots required for fitGAM.

Evaluate the optimal number of knots required for fitGAM.

evaluateK(counts, U = NULL, pseudotime, cellWeights, nGenes = 500,
  k = 3:10, weights = NULL, seed = 81, offset = NULL, ncores = 2,
  aicDiff = 2, bicDiff = 2)

Arguments

counts
the count matrix.

U
the design matrix of fixed effects. The design matrix should not contain an intercept to ensure identifiability.

pseudotime
a matrix of pseudotime values, each row represents a cell and each column represents a lineage.

cellWeights
a matrix of cell weights defining the probability that a cell belongs to a particular lineage. Each row represents a cell and each column represents a lineage.

nGenes
The number of genes to use in the evaluation. Genes will be randomly selected. 500 by default.

k
The range of knots to evaluate. `3:10` by default.

weights
Optional: a matrix of weights with identical dimensions as the counts matrix. Usually a matrix of zero-inflation weights.

seed
Optional: the seed used for assigning cells to lineages.

offset
Optional: the offset, on log-scale. If NULL, TMM is used to account for differences in sequencing depth, see fitGAM.

ncores
Number of cores to use.

aicDiff
Used for selecting genes with significantly varying AIC values over the range of evaluated knots to make the barplot output. Default is set to 2, meaning that only genes whose AIC range is larger than 2 will be used to check for the optimal number of knots through the barplot visualization that is part of the output of this function.

bicDiff
Used for selecting genes with significantly varying BIC values over the range of evaluated knots to make the barplot output. Default is set to 2, meaning that only genes whose BIC range is larger than 2 will be used to check for the optimal number of knots through the barplot visualization that is part of the output of this function. edgeR::calcNormFactors. Alternatively, this may also be a matrix of the same dimensions as the expression matrix.

Value

A plot of average AIC value over the range of selected knots, and a matrix of AIC values for the selected genes (rows) and the range of knots (columns).

Examples

# NOT RUN {
set.seed(8)
download.file("https://github.com/statOmics/tradeSeqPaper/raw/master/data/se_paul.rda",
destfile="./se_paul.rda")
load("./se_paul.rda")
se <- se[( 20:31)[-7], 25:40]
pseudotimes <- matrix(runif(ncol(se) * 2, 0, 5), ncol = 2)
cellWeights <- matrix(runif(ncol(se) * 2, 0, 1), ncol = 2)
gamList <- fitGAM(counts = as.matrix(
                      SummarizedExperiment::assays(se)$counts),
                  pseudotime = pseudotimes, cellWeights = cellWeights,
                  nknots = 5)
aicK <- evaluateK(counts = as.matrix(
                      SummarizedExperiment::assays(se)$counts),
                  pseudotime = pseudotimes, cellWeights = cellWeights,
                  nGenes=100, k=3:5, ncores=2)
# }