Get smoother p-value as returned by mgcv.

Return smoother p-values from the mgcv package.

getSmootherPvalues(models)

Arguments

models
the GAM models, typically the output from fitGAM. Note that this function only works when models is a list.

Value

a matrix with the p-value associated with each lineage's smoother. The matrix has one row per gene where the fitting procedure converged.

Examples

data(gamList, package = "tradeSeq")
getSmootherPvalues(gamList)

#>                s(t1):l1  s(t2):l2
#> 1700012B09Rik 1.0000000 1.0000000
#> 1700017B05Rik 0.1989893 0.6537023
#> 1700020L24Rik 1.0000000 1.0000000
#> 1810009A15Rik 0.9462776 0.4699110
#> 1810011H11Rik 1.0000000 1.0000000
#> 1810027O10Rik 0.5469890 0.5261559
#> 1810031K17Rik 0.1953402 0.3535604
#> 1810033B17Rik 1.0000000 1.0000000
#> 1810037I17Rik 0.4508156 0.4812300
#> 2010002N04Rik 1.0000000 1.0000000
#> 2010011I20Rik 0.9131269 0.8924305