predictSmooth

Get smoothers estimated by tradeSeq along a grid. This function does not return fitted values but rather the predicted mean smoother, for a user-defined grid of points.

predictSmooth(models, ...)

# S4 method for SingleCellExperiment
predictSmooth(models, gene, nPoints = 100, tidy = TRUE)

# S4 method for list
predictSmooth(models, gene, nPoints = 100)

Arguments

models
Either the SingleCellExperiment object obtained after running fitGAM, or the specific GAM model for the corresponding gene, if working with the list output of tradeSeq.

...
parameters including:

gene
Either a vector of gene names or an integer vector, corresponding to the row(s) of the gene(s).

nPoints
The number of points used to create the grid along the smoother for each lineage. Defaults to 100.

tidy
Logical: return tidy output. If TRUE, returns a data.frame specifying lineage, gene, pseudotime and value of estimated smoother. If FALSE, returns matrix of predicted smoother values, where each row is a gene and each column is a point on the uniform grid along the lineage. For example, if the trajectory consists of 2 lineages and nPoints=100, then the returned matrix will have 2*100 columns, where the first 100 correspond to the first lineage and columns 101-200 to the second lineage.

Value

A matrix with estimated averages.

Examples

data(gamList, package = "tradeSeq")
predictSmooth(models = gamList, gene = 1)

#>     lineage1_1   lineage1_2   lineage1_3  lineage1_4  lineage1_5   lineage1_6
#> 1 6.787343e-18 6.848751e-18 6.910715e-18 6.97324e-18 7.03633e-18 7.099992e-18
#>     lineage1_7   lineage1_8   lineage1_9  lineage1_10  lineage1_11  lineage1_12
#> 1 7.164229e-18 7.229047e-18 7.294452e-18 7.360448e-18 7.427042e-18 7.494238e-18
#>    lineage1_13  lineage1_14  lineage1_15  lineage1_16  lineage1_17  lineage1_18
#> 1 7.562042e-18 7.630459e-18 7.699496e-18 7.769157e-18 7.839448e-18 7.910376e-18
#>    lineage1_19  lineage1_20  lineage1_21 lineage1_22  lineage1_23  lineage1_24
#> 1 7.981945e-18 8.054161e-18 8.127031e-18 8.20056e-18 8.274755e-18 8.349621e-18
#>    lineage1_25 lineage1_26  lineage1_27  lineage1_28  lineage1_29  lineage1_30
#> 1 8.425164e-18 8.50139e-18 8.578307e-18 8.655919e-18 8.734233e-18 8.813256e-18
#>    lineage1_31  lineage1_32 lineage1_33  lineage1_34  lineage1_35  lineage1_36
#> 1 8.892994e-18 8.973453e-18 9.05464e-18 9.136562e-18 9.219225e-18 9.302636e-18
#>    lineage1_37  lineage1_38  lineage1_39  lineage1_40  lineage1_41 lineage1_42
#> 1 9.386801e-18 9.471728e-18 9.557424e-18 9.643894e-18 9.731147e-18 9.81919e-18
#>    lineage1_43  lineage1_44  lineage1_45 lineage1_46 lineage1_47  lineage1_48
#> 1 9.908029e-18 9.997672e-18 1.008813e-17 1.01794e-17 1.02715e-17 1.036443e-17
#>   lineage1_49  lineage1_50 lineage1_51  lineage1_52  lineage1_53  lineage1_54
#> 1 1.04582e-17 1.055282e-17 1.06483e-17 1.074464e-17 1.084185e-17 1.093994e-17
#>    lineage1_55  lineage1_56  lineage1_57  lineage1_58  lineage1_59  lineage1_60
#> 1 1.103892e-17 1.113879e-17 1.123957e-17 1.134126e-17 1.144387e-17 1.154741e-17
#>    lineage1_61  lineage1_62  lineage1_63  lineage1_64  lineage1_65  lineage1_66
#> 1 1.165188e-17 1.175731e-17 1.186368e-17 1.197102e-17 1.207932e-17 1.218861e-17
#>    lineage1_67  lineage1_68  lineage1_69  lineage1_70  lineage1_71  lineage1_72
#> 1 1.229889e-17 1.241016e-17 1.252244e-17 1.263574e-17 1.275006e-17 1.286542e-17
#>    lineage1_73  lineage1_74  lineage1_75  lineage1_76  lineage1_77 lineage1_78
#> 1 1.298182e-17 1.309927e-17 1.321778e-17 1.333737e-17 1.345804e-17 1.35798e-17
#>    lineage1_79  lineage1_80  lineage1_81  lineage1_82  lineage1_83  lineage1_84
#> 1 1.370267e-17 1.382664e-17 1.395174e-17 1.407797e-17 1.420534e-17 1.433386e-17
#>    lineage1_85 lineage1_86  lineage1_87  lineage1_88  lineage1_89  lineage1_90
#> 1 1.446354e-17 1.45944e-17 1.472645e-17 1.485968e-17 1.499413e-17 1.512979e-17
#>    lineage1_91 lineage1_92  lineage1_93  lineage1_94  lineage1_95  lineage1_96
#> 1 1.526667e-17 1.54048e-17 1.554417e-17 1.568481e-17 1.582672e-17 1.596991e-17
#>   lineage1_97  lineage1_98 lineage1_99 lineage1_100   lineage2_1   lineage2_2
#> 1 1.61144e-17 1.626019e-17 1.64073e-17 1.655575e-17 1.196904e-17 1.195194e-17
#>     lineage2_3  lineage2_4   lineage2_5   lineage2_6   lineage2_7   lineage2_8
#> 1 1.193485e-17 1.19178e-17 1.190077e-17 1.188376e-17 1.186678e-17 1.184982e-17
#>     lineage2_9  lineage2_10  lineage2_11  lineage2_12  lineage2_13  lineage2_14
#> 1 1.183288e-17 1.181597e-17 1.179909e-17 1.178223e-17 1.176539e-17 1.174857e-17
#>    lineage2_15  lineage2_16  lineage2_17  lineage2_18  lineage2_19  lineage2_20
#> 1 1.173178e-17 1.171502e-17 1.169828e-17 1.168156e-17 1.166486e-17 1.164819e-17
#>    lineage2_21  lineage2_22  lineage2_23  lineage2_24 lineage2_25  lineage2_26
#> 1 1.163155e-17 1.161493e-17 1.159833e-17 1.158175e-17 1.15652e-17 1.154867e-17
#>    lineage2_27  lineage2_28  lineage2_29 lineage2_30  lineage2_31 lineage2_32
#> 1 1.153217e-17 1.151569e-17 1.149923e-17 1.14828e-17 1.146639e-17   1.145e-17
#>    lineage2_33 lineage2_34  lineage2_35  lineage2_36  lineage2_37  lineage2_38
#> 1 1.143364e-17 1.14173e-17 1.140098e-17 1.138469e-17 1.136842e-17 1.135217e-17
#>    lineage2_39  lineage2_40  lineage2_41  lineage2_42  lineage2_43  lineage2_44
#> 1 1.133595e-17 1.131975e-17 1.130357e-17 1.128742e-17 1.127129e-17 1.125518e-17
#>    lineage2_45  lineage2_46  lineage2_47  lineage2_48  lineage2_49  lineage2_50
#> 1 1.123909e-17 1.122303e-17 1.120699e-17 1.119098e-17 1.117499e-17 1.115902e-17
#>    lineage2_51  lineage2_52  lineage2_53  lineage2_54  lineage2_55  lineage2_56
#> 1 1.114307e-17 1.112714e-17 1.111124e-17 1.109536e-17 1.107951e-17 1.106367e-17
#>    lineage2_57  lineage2_58  lineage2_59  lineage2_60  lineage2_61  lineage2_62
#> 1 1.104786e-17 1.103207e-17 1.101631e-17 1.100056e-17 1.098484e-17 1.096914e-17
#>    lineage2_63  lineage2_64  lineage2_65  lineage2_66  lineage2_67  lineage2_68
#> 1 1.095347e-17 1.093782e-17 1.092218e-17 1.090658e-17 1.089099e-17 1.087542e-17
#>    lineage2_69  lineage2_70  lineage2_71  lineage2_72  lineage2_73  lineage2_74
#> 1 1.085988e-17 1.084436e-17 1.082887e-17 1.081339e-17 1.079794e-17 1.078251e-17
#>   lineage2_75  lineage2_76  lineage2_77 lineage2_78  lineage2_79  lineage2_80
#> 1 1.07671e-17 1.075171e-17 1.073634e-17  1.0721e-17 1.070568e-17 1.069038e-17
#>   lineage2_81  lineage2_82  lineage2_83 lineage2_84  lineage2_85  lineage2_86
#> 1 1.06751e-17 1.065985e-17 1.064461e-17 1.06294e-17 1.061421e-17 1.059904e-17
#>    lineage2_87  lineage2_88  lineage2_89  lineage2_90  lineage2_91  lineage2_92
#> 1 1.058389e-17 1.056877e-17 1.055367e-17 1.053858e-17 1.052352e-17 1.050848e-17
#>    lineage2_93  lineage2_94 lineage2_95  lineage2_96  lineage2_97 lineage2_98
#> 1 1.049347e-17 1.047847e-17 1.04635e-17 1.044854e-17 1.043361e-17 1.04187e-17
#>    lineage2_99 lineage2_100
#> 1 1.040381e-17 1.038894e-17