Perform statistical test to check for DE between starting point and the end stages of every lineage

Perform statistical test to check for DE between starting point and the end stages of every lineage

startVsEndTest(models, ...)

# S4 method for SingleCellExperiment
startVsEndTest(models, global = TRUE,
  lineages = FALSE, pseudotimeValues = NULL)

# S4 method for list
startVsEndTest(models, global = TRUE,
  lineages = FALSE, pseudotimeValues = NULL)

Arguments

models
the list of GAMs, typically the output from fitGAM.

...
parameters including:

global
If TRUE, test for all lineages simultaneously.

lineages
If TRUE, test for all lineages independently.

pseudotimeValues
a vector of length 2, specifying two pseudotime values to be compared against each other, for every lineage of the trajectory. @details Note that this test assumes that all lineages start at a pseudotime value of zero, which is the starting point against which the end point is compared.

Value

A matrix with the wald statistic, the number of df and the p-value associated with each gene for all the tests performed. If the testing procedure was unsuccessful, the procedure will return NA test statistics and p-values. If both global and lineages are TRUE, then a matrix of p-values is returned.

Examples

data(gamList, package = "tradeSeq")
startVsEndTest(gamList, global = TRUE, lineages = TRUE)

#>                     global  lineage1  lineage2
#> 1700012B09Rik 1.489419e-16 1.0000000 1.0000000
#> 1700017B05Rik 4.316466e-01 0.5222756 0.8814257
#> 1700020L24Rik 1.489419e-16 1.0000000 1.0000000
#> 1810009A15Rik 1.384770e+00 0.9041920 0.2417635
#> 1810011H11Rik 1.489419e-16 1.0000000 1.0000000
#> 1810027O10Rik 2.589421e+00 0.4675587 0.1510435
#> 1810031K17Rik 1.805065e+00 0.2046942 0.6575575
#> 1810033B17Rik 1.489419e-16 1.0000000 1.0000000
#> 1810037I17Rik 7.002206e-01 0.4508052 0.7167850
#> 2010002N04Rik 1.489419e-16 1.0000000 1.0000000
#> 2010011I20Rik 4.511815e-01 0.5078248 0.9105017