Get smoother Chi-squared test statistics.

Return test statistics from the mgcv package.

getSmootherTestStats(models)

Arguments

models
the GAM models, typically the output from fitGAM. Note that this function only works when models is a list.

Value

a matrix with the wald statistics associated with each lineage's smoother. The matrix has one row per gene where the fitting procedure converged.

Examples

data(gamList, package = "tradeSeq")
getSmootherPvalues(gamList)

#>                s(t1):l1  s(t2):l2
#> 1700012B09Rik 1.0000000 1.0000000
#> 1700017B05Rik 0.1994792 0.6518374
#> 1700020L24Rik 1.0000000 1.0000000
#> 1810009A15Rik 0.8958601 0.4692757
#> 1810011H11Rik 1.0000000 1.0000000
#> 1810027O10Rik 0.5420688 0.5261227
#> 1810031K17Rik 0.1956860 0.3535415
#> 1810033B17Rik 1.0000000 1.0000000
#> 1810037I17Rik 0.4508141 0.4812285
#> 2010002N04Rik 1.0000000 1.0000000
#> 2010011I20Rik 0.9087768 0.8879531