Assess differential expression pattern between lineages.

Assess differential expression pattern between lineages.

# S4 method for list
patternTest(models, global = TRUE, pairwise = FALSE,
  nPoints = 100)

# S4 method for SingleCellExperiment
patternTest(models, global = TRUE,
  pairwise = FALSE, nPoints = 100)

Arguments

models
the list of GAMs, typically the output from fitGAM.

global
If TRUE, test for all pairwise comparisons simultaneously.

pairwise
If TRUE, test for all pairwise comparisons independently.

nPoints
the number of points to be compared between lineages.

Value

A matrix with the wald statistic, the number of df and the p-value associated with each gene for all the tests performed. If the testing procedure was unsuccessful, the procedure will return NA test statistics and p-values.

Examples

data(gamList, package = "tradeSeq")
patternTest(gamList, global = TRUE, pairwise = TRUE)

#> Only two lineages; skipping pairwise comparison.

#>                   waldStat df    pvalue
#> 1700012B09Rik 9.664503e-17  2 1.0000000
#> 1700017B05Rik 6.729675e+00  8 0.5660597
#> 1700020L24Rik 9.664503e-17  2 1.0000000
#> 1810009A15Rik 1.859652e+00  8 0.9850163
#> 1810011H11Rik 9.664503e-17  2 1.0000000
#> 1810027O10Rik 1.655389e+00  8 0.9898174
#> 1810031K17Rik 5.841343e+00  8 0.6649984
#> 1810033B17Rik 9.664503e-17  2 1.0000000
#> 1810037I17Rik 1.486465e+00  5 0.9146265
#> 2010002N04Rik 9.664503e-17  2 1.0000000
#> 2010011I20Rik 1.852404e+00  7 0.9675559